renamed interfaces
