+++ hardware/dummies/spectroscopy_camera_dummy.py
